+++ frontend/src/pages/AdvancedSearchPage.tsx
@@ -20,6 +20,7 @@ import { Link } from "react-router-dom";
 import { useAsync } from "react-use";
 
 import { aironeApiClientV2 } from "../apiclient/AironeApiClientV2";
+import { fuzzyMatch } from "../utils/StringUtil";
 
 import { advancedSearchResultPath, topPath } from "Routes";
 import { AironeBreadcrumbs } from "components/common/AironeBreadcrumbs";
@@ -111,9 +112,7 @@ export const AdvancedSearchPage: FC = () => {
         entities.value
           .filter(
             (entity) =>
-              entity.name
-                .toLowerCase()
-                .indexOf(entityNameFilter.toLowerCase()) !== -1 &&
+              fuzzyMatch(entity.name, entityNameFilter) &&
               selectedEntityIds.indexOf(entity.id) === -1
           )
           .map((entity) => toggleSelectedEntityIds(entity.id))
@@ -126,7 +125,7 @@ export const AdvancedSearchPage: FC = () => {
         attrs.value
           .filter(
             (attr) =>
-              attr.toLowerCase().indexOf(attrNameFilter.toLowerCase()) !== -1 &&
+              fuzzyMatch(attr, attrNameFilter) &&
               selectedAttrs.indexOf(attr) === -1
           )
           .map((attr) => toggleSelectedAttrs(attr))
@@ -158,12 +157,7 @@ export const AdvancedSearchPage: FC = () => {
           {!entities.loading ? (
             <>
               {entities.value
-                .filter(
-                  (entity) =>
-                    entity.name
-                      .toLowerCase()
-                      .indexOf(entityNameFilter.toLowerCase()) !== -1
-                )
+                .filter((entity) => fuzzyMatch(entity.name, entityNameFilter))
                 .map((entity) => (
                   <ListItem
                     key={entity.id}
@@ -207,11 +201,7 @@ export const AdvancedSearchPage: FC = () => {
           {!attrs.loading ? (
             <>
               {attrs.value
-                .filter(
-                  (attr) =>
-                    attr.toLowerCase().indexOf(attrNameFilter.toLowerCase()) !==
-                    -1
-                )
+                .filter((attr) => fuzzyMatch(attr, attrNameFilter))
                 .map((attr) => (
                   <ListItem
                     key={attr}