+++ role/api_v1/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('<int:role_id>/', views.RoleAPI.as_view()),
+]