+++ role/settings.py
@@ -1,0 +1,5 @@
+from airone.lib.settings import Settings
+
+CONFIG = Settings({
+    'MAX_LIST_ROLE': 20,
+})