+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 * Added add and remove attributes depending on entity in get entry api v2
   Contributed by @hinashi
+* Added custom processing when after delete entry
+  Contributed by @hinashi
 
 ### Changed